bundle/run: move runPageURL into libs/workspaceurls
The legacy-to-path run URL conversion is needed outside `bundle run`, so it
moves to libs/workspaceurls next to JobRunPath, which it already used.

Pure move: the conversion, its doc comment and its tests are unchanged.
